made a not hacky, completly portible fix for the box shadows!
tldr: we colone the rect/div with the box shadow, apply a filter on it as before but with a feComposite with operator to out. (this doesn't fill the shape itself, only the shadow outside of it)

Rect fill doesn't mater now, so we just use the shadow colour. Insert the clone before the real rect.